refactor sidebar, add active class and use next navigation
